pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "uvicorn>=0.42.0",
]
